Refactored Testone to create Testtwo
